refactor: 공통 훅 폴더구조 분리
