+++ frontend/src/pages/manager/spaceHome/SpaceHome.tsx
@@ -8,8 +8,8 @@ import FloatingIconButton from '../../../components/@common/buttons/floatingIcon
 import ImageGrid from '../../../components/@common/imageGrid/ImageGrid';
 import SpaceHeader from '../../../components/spaceHeader/SpaceHeader';
 import { INFORMATION } from '../../../constants/messages';
+import useIntersectionObserver from '../../../hooks/@common/useIntersectionObserver';
 import useDownload from '../../../hooks/useDownload';
-import useIntersectionObserver from '../../../hooks/useIntersectionObserver';
 import usePhotosBySpaceCode from '../../../hooks/usePhotosBySpaceCode';
 import { theme } from '../../../styles/theme';
 import { goToTop } from '../../../utils/goToTop';
